:hammer: Support auto-running of seeds on app boot.

+++ api/src/data/seeds/002_centre_submissions.ts
@@ -1,6 +1,10 @@
 import type * as knex from "knex"
 
 exports.seed = async function (knex: knex.Knex, Promise: any) {
+  const hasCenterSubmissionsTable = await knex.schema.hasTable('centre_submissions')
+
+  if (!hasCenterSubmissionsTable) return
+
   await knex("centre_submissions")
     .truncate()
     .then(function () {
